Refactor about detecting the RenderPipeline
