+++ Assets/UniGLTF/Runtime/UniGLTF/IO/MaterialIO/Import/MaterialDescriptorGeneratorUtility.cs
@@ -1,0 +1,16 @@
+namespace UniGLTF
+{
+    public static class MaterialDescriptorGeneratorUtility
+    {
+        public static IMaterialDescriptorGenerator GetValidGltfMaterialDescriptorGenerator()
+        {
+            return RenderPipelineUtility.GetRenderPipelineType() switch
+            {
+                RenderPipelineTypes.UniversalRenderPipeline => new UrpGltfMaterialDescriptorGenerator(),
+                RenderPipelineTypes.BuiltinRenderPipeline => new BuiltInGltfMaterialDescriptorGenerator(),
+                _ => new BuiltInGltfMaterialDescriptorGenerator(),
+            };
+        }
+    }
+}
+
