+++ Assets/VRM10/Runtime/IO/Material/URP/Import/Vrm10MaterialDescriptorGeneratorDescriptorUtility.cs
@@ -6,7 +6,7 @@ namespace UniVRM10
     {
         public static IMaterialDescriptorGenerator GetValidVrm10MaterialDescriptorGenerator()
         {
-            return MaterialDescriptorGeneratorUtility.GetRenderPipelineType() switch
+            return RenderPipelineUtility.GetRenderPipelineType() switch
             {
                 RenderPipelineTypes.UniversalRenderPipeline => new UrpVrm10MaterialDescriptorGenerator(),
                 RenderPipelineTypes.BuiltinRenderPipeline => new BuiltInVrm10MaterialDescriptorGenerator(),
